# apps/frontend/src/api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/monitoreo-ruta/monitoreo/{despachoId}": {
      "get": {
        "description": "Endpoint SSE (Server-Sent Events) para recibir actualizaciones en tiempo real de un despacho específico.",
        "operationId": "MonitoreoEnRutaController_monitoreoEnTiempoReal",
        "parameters": [
          {
            "name": "despachoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conexión SSE establecida correctamente",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "type",
                    "id",
                    "retry"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MonitoreoDto"
                    },
                    "type": {
                      "type": "string",
                      "example": "monitoreo"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "number",
                      "example": 5000
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Suscripción a datos de monitoreo",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/despachos": {
      "get": {
        "description": "Obtiene los despachos paginados con ordenamiento por fecha de creación",
        "operationId": "MonitoreoEnRutaController_getDespachosPaginados",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Despachos obtenidos correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "items",
                        "meta"
                      ],
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/DespachoDto"
                          }
                        },
                        "meta": {
                          "type": "object",
                          "required": [
                            "total",
                            "page",
                            "limit",
                            "totalPages",
                            "hasNextPage",
                            "hasPreviousPage"
                          ],
                          "properties": {
                            "total": {
                              "type": "number",
                              "example": 100
                            },
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPages": {
                              "type": "number",
                              "example": 10
                            },
                            "hasNextPage": {
                              "type": "boolean",
                              "example": false
                            },
                            "hasPreviousPage": {
                              "type": "boolean",
                              "example": false
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener despachos con paginación",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/ultimo-monitoreo/{despachoId}": {
      "get": {
        "description": "Obtiene el último monitoreo de un despacho",
        "operationId": "MonitoreoEnRutaController_getUltimoMonitoreo",
        "parameters": [
          {
            "name": "despachoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Último monitoreo obtenido correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UltimoMonitoreoDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontró el despacho o no tiene monitoreos registrados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener el último monitoreo de un despacho",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/incidencia": {
      "post": {
        "description": "Crea una incidencia",
        "operationId": "MonitoreoEnRutaController_createIncidencia",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidenciaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Incidencia creada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crear una incidencia",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/incidencia/{incidenciaId}": {
      "patch": {
        "description": "Actualiza el estado de una incidencia",
        "operationId": "MonitoreoEnRutaController_updateIncidenciaState",
        "parameters": [
          {
            "name": "incidenciaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncidenciaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Incidencia actualizada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Actualizar el estado de una incidencia",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/solucionar-incidencia/{incidenciaId}": {
      "post": {
        "description": "Soluciona una incidencia",
        "operationId": "MonitoreoEnRutaController_solucionarIncidencia",
        "parameters": [
          {
            "name": "incidenciaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolucionIncidenciaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Incidencia solucionada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Solucionar una incidencia",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/incidencias": {
      "get": {
        "description": "Obtiene las incidencias paginadas con ordenamiento por fecha de registro",
        "operationId": "MonitoreoEnRutaController_getIncidenciasPaginadas",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Incidencias obtenidas correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "items",
                        "meta"
                      ],
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/IncidenciaDto"
                          }
                        },
                        "meta": {
                          "type": "object",
                          "required": [
                            "total",
                            "page",
                            "limit",
                            "totalPages",
                            "hasNextPage",
                            "hasPreviousPage"
                          ],
                          "properties": {
                            "total": {
                              "type": "number",
                              "example": 100
                            },
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPages": {
                              "type": "number",
                              "example": 10
                            },
                            "hasNextPage": {
                              "type": "boolean",
                              "example": false
                            },
                            "hasPreviousPage": {
                              "type": "boolean",
                              "example": false
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener incidencias con paginación",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/despachos/{despachoId}": {
      "get": {
        "description": "Obtiene los datos básicos de un despacho específico",
        "operationId": "MonitoreoEnRutaController_getDespachoDatos",
        "parameters": [
          {
            "name": "despachoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos básicos del despacho obtenidos correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DespachoDatosDto"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener datos básicos de un despacho",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/despachos/{despachoId}/puntos-ruta": {
      "get": {
        "description": "Obtiene los puntos de la ruta de un despacho específico",
        "operationId": "MonitoreoEnRutaController_getPuntosRuta",
        "parameters": [
          {
            "name": "despachoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Puntos de la ruta obtenidos correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PuntoRutaDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener puntos de la ruta de un despacho",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/despachos/{despachoId}/puntos-visitados": {
      "get": {
        "description": "Obtiene los puntos visitados de un despacho específico",
        "operationId": "MonitoreoEnRutaController_getPuntosVisitados",
        "parameters": [
          {
            "name": "despachoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Puntos visitados obtenidos correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PuntoVisitadoDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener puntos visitados de un despacho",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
    "/api/monitoreo-ruta/incidencias/{incidenciaId}": {
      "get": {
        "description": "Obtiene una incidencia específica",
        "operationId": "MonitoreoEnRutaController_getIncidenciaDatos",
        "parameters": [
          {
            "name": "incidenciaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Incidencia obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/IncidenciaDatosDto"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener una incidencia",
        "tags": [
          "Monitoreo en Ruta"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Transporte",
    "description": "API para el sistema de monitoreo de transporte",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Ingresa tu token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "UltimoMonitoreoDto": {
        "type": "object",
        "properties": {
          "id_monitoreo_ruta": {
            "type": "string",
            "description": "ID del monitoreo de ruta",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "ubicacion_actual": {
            "type": "object",
            "description": "Ubicación actual del vehículo",
            "example": {
              "x": -77.0428,
              "y": -12.0464
            }
          }
        },
        "required": [
          "id_monitoreo_ruta",
          "ubicacion_actual"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "description": "Estado de la respuesta",
            "example": false,
            "nullable": false
          },
          "message": {
            "type": "string",
            "description": "Mensaje de error",
            "example": "Error interno del servidor",
            "nullable": false
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
...
    }
  }
}